Fetch weather info from API

+++ app/src/main/java/com/github/pedrodimoura/easyweather/today/ui/model/TodayWeatherInformation.kt
@@ -2,7 +2,7 @@ package com.github.pedrodimoura.easyweather.today.ui.model
 
 data class TodayWeatherInformation(
     val temperature: String,
-    val rawTemperature: Int,
+    val rawTemperature: Double,
     val dateTime: String,
     val location: String,
 )
